move PersonalMobileDevicesPage
