+++ frontend/src/employee-frontend/App.tsx
@@ -23,7 +23,7 @@ import { Footer } from './components/Footer'
 import Header from './components/Header'
 import LoginPage from './components/LoginPage'
 import MobilePairingModal from './components/MobilePairingModal'
-import PersonalMobileDevicesPage from './components/PersonalMobileDevicesPage'
+import PersonalMobileDevicesPage from './components/personal-mobile-devices/PersonalMobileDevicesPage'
 import Reports from './components/Reports'
 import UnitPage from './components/UnitPage'
 import Units from './components/Units'
