+++ frontend/src/employee-frontend/components/personal-mobile-devices/PersonalMobileDevicesPage.tsx
@@ -25,12 +25,12 @@ import {
   deleteMobileDevice,
   getPersonalMobileDevices,
   putMobileDeviceName
-} from '../generated/api-clients/pairing'
-import { useTranslation } from '../state/i18n'
-import { UIContext } from '../state/ui'
-import { UserContext } from '../state/user'
+} from '../../generated/api-clients/pairing'
+import { useTranslation } from '../../state/i18n'
+import { UIContext } from '../../state/ui'
+import { UserContext } from '../../state/user'
 
-import { renderResult } from './async-rendering'
+import { renderResult } from '../async-rendering'
 
 const getPersonalMobileDevicesResult = wrapResult(getPersonalMobileDevices)
 const deleteMobileDeviceResult = wrapResult(deleteMobileDevice)
